添加文章评论的插入

+++ src/main/java/cn/LJW/Entities/article/CommentBean.kt
@@ -1,11 +1,13 @@
 package cn.LJW.Entities.article
 
+import java.math.BigInteger
+
 data class CommentBean(
-    val synergyId: Long,
+    val synergyId: BigInteger,
     val paragraphId: Long,
     val userId: String,
     val commentId: Long,
     val cotent: String,
     val favor: Long,
-    val createTime: Long
+    val createTime: BigInteger
 )
